feat: Add File component for displaying document information
This commit adds a new component called File, which is responsible for displaying information about a document. It imports the necessary dependencies, including the DocumentIcon image from the /public/images directory. The File component receives props for the document title and file name, and renders them along with the document icon. This component is intended to be used in other parts of the application to display document information.